refactor: DTO Response, Request 분리
