+++ src/main/java/com/gamemoonchul/application/MemberService.java
@@ -6,7 +6,7 @@ import com.gamemoonchul.application.converter.MemberConverter;
 import com.gamemoonchul.domain.entity.Member;
 import com.gamemoonchul.domain.status.MemberStatus;
 import com.gamemoonchul.infrastructure.repository.MemberRepository;
-import com.gamemoonchul.infrastructure.web.dto.MemberResponseDto;
+import com.gamemoonchul.infrastructure.web.dto.response.MemberResponseDto;
 import jakarta.transaction.Transactional;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.NotEmpty;
