+++ src/main/java/com/gamemoonchul/application/PostService.java
@@ -9,8 +9,8 @@ import com.gamemoonchul.domain.entity.VoteOptions;
 import com.gamemoonchul.domain.status.PostStatus;
 import com.gamemoonchul.infrastructure.repository.PostRepository;
 import com.gamemoonchul.infrastructure.repository.VoteOptionRepository;
-import com.gamemoonchul.infrastructure.web.dto.PostResponseDto;
-import com.gamemoonchul.infrastructure.web.dto.PostUploadRequest;
+import com.gamemoonchul.infrastructure.web.dto.response.PostResponseDto;
+import com.gamemoonchul.infrastructure.web.dto.request.PostUploadRequest;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
