+++ src/main/java/com/gamemoonchul/application/converter/MemberConverter.java
@@ -1,12 +1,10 @@
 package com.gamemoonchul.application.converter;
 
-import com.gamemoonchul.config.apple.AppleCredential;
-import com.gamemoonchul.config.oauth.user.OAuth2Provider;
 import com.gamemoonchul.config.oauth.user.OAuth2UserInfo;
 import com.gamemoonchul.domain.entity.Member;
 import com.gamemoonchul.domain.entity.redis.RedisMember;
 import com.gamemoonchul.domain.enums.MemberRole;
-import com.gamemoonchul.infrastructure.web.dto.MemberResponseDto;
+import com.gamemoonchul.infrastructure.web.dto.response.MemberResponseDto;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
