+++ src/main/java/com/gamemoonchul/domain/model/dto/CommentSaveDto.java
@@ -1,10 +1,9 @@
 package com.gamemoonchul.domain.model.dto;
 
-import com.gamemoonchul.infrastructure.web.dto.CommentRequest;
 import lombok.Builder;
 
 @Builder
-public record CommentSaveDto (
+public record CommentSaveDto(
         Long parentId,
         String content,
         Long postId
