+++ src/main/java/com/gamemoonchul/infrastructure/web/MemberApiController.java
@@ -4,7 +4,7 @@ import com.gamemoonchul.application.MemberService;
 import com.gamemoonchul.common.annotation.MemberSession;
 import com.gamemoonchul.domain.entity.Member;
 import com.gamemoonchul.infrastructure.web.common.RestControllerWithEnvelopPattern;
-import com.gamemoonchul.infrastructure.web.dto.MemberResponseDto;
+import com.gamemoonchul.infrastructure.web.dto.response.MemberResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
