+++ src/main/java/com/gamemoonchul/infrastructure/web/PostOpenApiController.java
@@ -3,7 +3,7 @@ package com.gamemoonchul.infrastructure.web;
 import com.gamemoonchul.application.PostOpenApiService;
 import com.gamemoonchul.infrastructure.web.common.Pagination;
 import com.gamemoonchul.infrastructure.web.common.RestControllerWithEnvelopPattern;
-import com.gamemoonchul.infrastructure.web.dto.PostResponseDto;
+import com.gamemoonchul.infrastructure.web.dto.response.PostResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
