+++ src/test/java/com/gamemoonchul/infrastructure/web/dto/MatchGameResponseTest.java
@@ -3,9 +3,9 @@ package com.gamemoonchul.infrastructure.web.dto;
 import com.gamemoonchul.TestDataBase;
 import com.gamemoonchul.domain.entity.riot.MatchGame;
 import com.gamemoonchul.domain.entity.riot.MatchGameDummy;
+import com.gamemoonchul.infrastructure.web.dto.response.MatchGameResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -23,7 +23,9 @@ class MatchGameResponseTest extends TestDataBase {
         assertEquals(matchGame.getGameCreation(), matchGameResponse.getGameCreation());
         assertEquals(matchGame.getGameMode(), matchGameResponse.getGameMode());
         assertEquals(matchGame.getGameDuration(), matchGameResponse.getGameDuration());
-        assertEquals(matchGame.getMatchUsers().size(), matchGameResponse.getMatchUsers().size());
+        assertEquals(matchGame.getMatchUsers()
+                .size(), matchGameResponse.getMatchUsers()
+                .size());
     }
 
 }
